+++ sync-for-expenses/RELEASES.md
@@ -709,3 +709,13 @@ Based on:
 - [typescript v3.3.0] sync-for-expenses
 ### Releases
 - [NPM v3.3.0] https://www.npmjs.com/package/@codat/sync-for-expenses/v/3.3.0 - sync-for-expenses
+
+## 2023-10-30 01:27:19
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.4.0] sync-for-expenses
+### Releases
+- [NPM v3.4.0] https://www.npmjs.com/package/@codat/sync-for-expenses/v/3.4.0 - sync-for-expenses